Revert "mx-food-templates"
