+++ gatsby-node.js
@@ -8,9 +8,6 @@ const legalRoutes = require("./routes/legal-routes");
 const repartidoresRoutes = require("./routes/repartidores-routes");
 const faqsRoutes = require("./routes/faqs-routes");
 const redirects = require("./routes/redirects");
-const cityRoutes = require("./routes/food-city-routes");
-const municipalityRoutes = require("./routes/food-municipality-routes");
-const neighbourhoodRoutes = require("./routes/food-neighbourhood-routes");
 
 // Implement the Gatsby API “createPages”. This is called once the
 // each route will import a ContentType List from Contentful
@@ -30,7 +27,4 @@ exports.createPages = async ({
   await legalRoutes.init(graphql, createPage);
   await repartidoresRoutes.init(graphql, createPage);
   await faqsRoutes.init(graphql, createPage);
-  await cityRoutes.init(graphql, createPage); 
-  await municipalityRoutes.init(graphql, createPage); 
-  await neighbourhoodRoutes.init(graphql, createPage); 
 };
