+++ src/pages/mx/food/colonia/{contentfulNeighbourhood.slug}.js
@@ -1,18 +1,24 @@
 import React from "react";
 import { graphql } from "gatsby";
-import Layout from "../../components/Layout";
-import FoodCityHero from "../../components/sections/FoodCityHero";
-import FoodCityBannerCTA from "../../components/sections/FoodCityBannerCTA";
-import FoodCityBannerCTA2 from "../../components/sections/FoodCityBannerCTA2";
-import FoodCityBannerCTA3 from "../../components/sections/FoodCityBannerCTA3";
-import FoodCityRestaurantCTA from "../../components/sections/FoodCityRestaurantCTA";
-import FoodCityList from "../../components/sections/FoodCityList";
-import FoodAppDownloads from "../../components/sections/FoodAppDownloads";
+import Layout from "../../../../components/Layout";
+import FoodCityHero from "../../../../components/sections/FoodCityHero";
+import FoodCityBannerCTA from "../../../../components/sections/FoodCityBannerCTA";
+import FoodCityBannerCTA2 from "../../../../components/sections/FoodCityBannerCTA2";
+import FoodCityBannerCTA3 from "../../../../components/sections/FoodCityBannerCTA3";
+import FoodCityRestaurantCTA from "../../../../components/sections/FoodCityRestaurantCTA";
+import FoodNeighborhoodList from "../../../../components/sections/FoodNeighborhoodList";
+import FoodAppDownloads from "../../../../components/sections/FoodAppDownloads";
 
 const FoodCity = ({ data }) => {
   const images = data.allContentfulAsset.nodes;
-  const itemsList = data.allContentfulCity.nodes;
-  const name = data.contentfulCity.name;
+  const cities = data.contentfulNeighbourhood.municipality
+    ? data.contentfulNeighbourhood.municipality.neighbourhood
+    : [];
+  const cityName = data.contentfulNeighbourhood.city.name;
+  const citySlug = data.contentfulNeighbourhood.city.slug;
+  const municipalityName = data.contentfulNeighbourhood.municipality.name;
+  const municipalitySlug = data.contentfulNeighbourhood.municipality.slug;
+  const name = data.contentfulNeighbourhood.name;
 
   const foodHeroBgImage = images.filter((image) => {
     return image.title === "mx.FoodHero.bgImage";
@@ -29,17 +35,20 @@ const FoodCity = ({ data }) => {
   const foodDeliveryDownloadsImages = images.filter((image) => {
     return image.title.indexOf("mx.FoodDeliveryDownloads.image") !== -1;
   });
+
+  //! Is not a good solution needs refactor to be more flexible in other countries
   const customBreadcrumb = [
-    { 
-      link: `https://web.didiglobal.com/mx/food/`, 
-      text: "Food"
+    { link: `https://web.didiglobal.com/mx/food/`, text: "Food" },
+    {
+      link: `https://web.didiglobal.com/mx/food/ciudad/${citySlug}/`,
+      text: cityName,
     },
     {
-      link: "#", 
-      text: name
-    }
-  ]
-
+      link: `https://web.didiglobal.com/mx/food/colonia/${municipalitySlug}/`,
+      text: municipalityName,
+    },
+    { link: "#", text: name },
+  ];
   return (
     <Layout
       title={`Pide Comida a Domicilio  en ${name} CDMX`}
@@ -48,23 +57,23 @@ const FoodCity = ({ data }) => {
     >
       <FoodCityHero
         bgImage={foodHeroBgImage}
-        data={data.contentfulCity}
+        data={data.contentfulNeighbourhood}
       ></FoodCityHero>
       <FoodCityBannerCTA
-        data={data.contentfulCity}
+        data={data.contentfulNeighbourhood}
         image={foodBusinessCTAImage}
       ></FoodCityBannerCTA>
       <FoodCityRestaurantCTA
-        data={data.contentfulCity}
+        data={data.contentfulNeighbourhood}
         image={foodDeliveryCTAImage}
       ></FoodCityRestaurantCTA>
       <FoodCityBannerCTA2
-        data={data.contentfulCity}
+        data={data.contentfulNeighbourhood}
         image={foodDeliveryCTAImage}
       ></FoodCityBannerCTA2>
-      <FoodCityList data={itemsList}></FoodCityList>
+      <FoodNeighborhoodList data={cities}></FoodNeighborhoodList>
       <FoodCityBannerCTA3
-        data={data.contentfulCity}
+        data={data.contentfulNeighbourhood}
         image={foodCTA3Image}
       ></FoodCityBannerCTA3>
       <FoodAppDownloads images={foodDeliveryDownloadsImages}></FoodAppDownloads>
@@ -76,16 +85,21 @@ export default FoodCity;
 
 export const query = graphql`
   query ($id: String) {
-    allContentfulCity ( filter: {id: {ne: $id}, country: {code: {eq: "mx"}}}) {
-      nodes{
-        id
+    contentfulNeighbourhood(id: { eq: $id }) {
+      name
+      slug
+      city {
         name
         slug
       }
-    }
-    contentfulCity(id: { eq: $id }) {
-      name
-      slug
+      municipality {
+        name
+        slug
+        neighbourhood {
+          name
+          slug
+        }
+      }
     }
     allContentfulAsset(
       filter: {
